requirements_win.txt
arxiv==2.1.3
b2sdk==2.6.0
beautifulsoup4==4.12.3
cachetools==5.5.0
docling==2.14.0
flask==3.1.0
chromadb==0.5.18
google-api-core==2.24.0
google-api-python-client==2.156.0
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.2.14
langchain-google-genai
langchain-groq==0.2.2
langgraph==0.2.60
origamibot==2.3.6
protobuf==5.28.3
pydantic==2.10.3
pymongo==4.10.1
pyngrok==7.2.2
pypdf==5.1.0
pywin32==307
retry==0.9.2
pyinstaller==6.11.1
keyring==25.6.0
waitress==3.0.2
jinja2~=3.1.5
python-dotenv~=1.0.1
requests~=2.32.3
hvac~=2.3.0